requirements.txt
# ECOLANG Streamlit Frontend Requirements
# Simplified - No heavy ML dependencies (those run in Colab)

streamlit>=1.28.0
httpx>=0.24.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.7.0
